requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==42.0.8
dnspython==2.7.0
Flask==3.1.1
idna==3.10
impacket==0.12.0
itsdangerous==2.2.0
Jinja2==3.1.6
ldap3==2.9.1
ldapdomaindump==0.10.0
MarkupSafe==3.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pyOpenSSL==24.0.0
pyreadline3==3.5.4
python-nmap==0.7.1
PyYAML==6.0.2
requests==2.32.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
